胜利失败界面修改

+++ frameworks/runtime-src/Classes/unit/ElementData.h
@@ -36,7 +36,7 @@ public:
     float recover;
     float scale;
     
-    float view_range;
+    int view_range;
     
     std::string bullet_name;
